Look at Ljvf and Alpha constant effects on the crash estimates
- Compare e^alpha for both intersections and segments
- Compare Ljvf for both intersections and segments
- Compare volume for both intersections and segments

The main inconsistencies here are for Ljvf and e^alpha. Volume seems to be fairly consistent.

e^alpha inconsistencies:
- Graph e^alpha at intersections vs e^alpha at roadways
- Find and graph ratio between e^alpha at intersections and e^alpha at roadways (for the same mode, outcome, volume class, functional class)

e^alpha is definitely much higher - usually 10x (or more) larger for intersections than roadways.

Ljvf is also larger, but it is also in different units (count vs miles). But I wasn't sure if there could also be an inconsistency in how Ljvf is calculated, so looked into it further:
- Compare Ljvf with overall project length/count
- Compare total Ljvf (per project and location type) with total project length/count
- Find and graph the ratio between Ljvf length/count and total project length/count

It looks like there is something off with the Ljvf calculations. There is generally no change between the total number of intersections in the project and the Ljvf for intersections, but Ljvf for roadways is a very small fraction of the total project length - and it seems to almost always be one constant value around 0.000189 (?) The next step is to try to figure out when/how/why this is happening.