Manually calculate crashes for one project: segments
For **segments only:**
- Combined Ljvf, alpha constants, Vmj into one table
- New columns multiplying Ljvf * e^alpha and Vmj * e^alpha and crashes = Ljvf * Vmj * e^alpha
- Graphed these new estimated crashes and crashes/mile (Vmj*e^alpha)
Calculating manually, it looks like the crashes for segments are also in the hundreds/thousands? It seems like the tool is doing some other step in between or using some other data at some point because I am using the same data that it should be using and the same process that it should be using, but getting much higher numbers of crashes. Maybe this has to do with the Ljvf calculation again...?